@@ -26,13 +26,13 @@
     "test:watch": "yarn test --watch"
   },
   "dependencies": {
-    "@apollo/client": "^3.1.3",
+    "@apollo/client": "^3.2.1",
     "@react-native-community/async-storage": "1.12.0",
     "@react-native-community/masked-view": "^0.1.10",
-    "@react-native-community/netinfo": "5.9.6",
+    "@react-native-community/netinfo": "5.9.7",
     "@react-native-community/slider": "3.0.3",
-    "@react-navigation/native": "5.7.3",
-    "@react-navigation/stack": "5.9.0",
+    "@react-navigation/native": "5.7.4",
+    "@react-navigation/stack": "5.9.1",
     "apollo-cache-persist": "^0.1.1",
     "graphql": "15.3.0",
     "graphql-tag": "2.11.0",
@@ -44,7 +44,7 @@
     "react": "16.13.1",
     "react-native": "0.63.2",
     "react-native-animatable": "1.3.3",
-    "react-native-device-info": "5.6.5",
+    "react-native-device-info": "6.0.4",
     "react-native-fs": "2.16.6",
     "react-native-gesture-handler": "1.7.0",
     "react-native-google-cast": "github:pct-org/react-native-google-cast#7e409ca7c129d628d5c2902944790bc611b64aaa",
@@ -52,12 +52,12 @@
     "react-native-linear-gradient": "2.5.6",
     "react-native-markdown-renderer": "^3.2.8",
     "react-native-orientation": "^3.1.3",
-    "react-native-reanimated": "^1.13.0",
-    "react-native-safe-area-context": "3.1.7",
-    "react-native-screens": "2.10.1",
+    "react-native-reanimated": "1.13.0",
+    "react-native-safe-area-context": "3.1.8",
+    "react-native-screens": "2.11.0",
     "react-native-splash-screen": "3.2.0",
     "react-native-static-server": "0.5.0",
-    "react-native-vector-icons": "7.0.0",
+    "react-native-vector-icons": "7.1.0",
     "react-navigation-collapsible": "5.6.4",
     "reanimated-bottom-sheet": "^1.0.0-alpha.22",
     "redux-persist-fs-storage": "^1.3.0",
@@ -65,36 +65,36 @@
     "update-react-native-app": "^1.3.1"
   },
   "devDependencies": {
-    "@babel/core": "7.11.4",
+    "@babel/core": "7.11.6",
     "@babel/plugin-proposal-class-properties": "7.10.4",
     "@babel/plugin-proposal-decorators": "7.10.5",
     "@babel/plugin-proposal-export-namespace-from": "7.10.4",
     "@babel/plugin-proposal-object-rest-spread": "7.11.0",
     "@babel/plugin-transform-flow-strip-types": "7.10.4",
-    "@babel/plugin-transform-runtime": "7.11.0",
+    "@babel/plugin-transform-runtime": "7.11.5",
     "@babel/preset-react": "7.10.4",
     "@babel/runtime": "7.11.2",
-    "@commitlint/cli": "9.1.2",
-    "@commitlint/config-conventional": "9.1.2",
+    "@commitlint/cli": "11.0.0",
+    "@commitlint/config-conventional": "11.0.0",
     "babel-eslint": "10.1.0",
     "babel-plugin-inline-import": "^3.0.0",
     "babel-plugin-module-resolver": "4.0.0",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "1.15.3",
-    "eslint": "7.7.0",
+    "enzyme-adapter-react-16": "1.15.5",
+    "eslint": "7.9.0",
     "eslint-config-airbnb": "18.2.0",
     "eslint-config-standard": "14.1.1",
     "eslint-config-standard-react": "9.2.0",
     "eslint-plugin-babel": "5.3.1",
     "eslint-plugin-import": "2.22.0",
-    "eslint-plugin-jest": "23.20.0",
+    "eslint-plugin-jest": "24.0.2",
     "eslint-plugin-jsx-a11y": "6.3.1",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-react": "7.20.6",
+    "eslint-plugin-react": "7.21.1",
     "eslint-plugin-standard": "4.0.1",
     "get-dev-paths": "^0.1.1",
-    "husky": "4.2.5",
+    "husky": "4.3.0",
     "jest": "26.4.2",
     "jest-enzyme": "^7.1.2",
     "jest-react-native": "^18.0.0",
