@@ -1,13 +1,12 @@
 import 'react-native-gesture-handler'
-import { AppRegistry, LogBox, YellowBox } from 'react-native'
+import { AppRegistry, LogBox } from 'react-native'
 import { enableScreens } from 'react-native-screens'
 
 enableScreens()
 
 LogBox.ignoreLogs([
   'Debugger and device times',
-  'Cache data may be lost when replacing',
-  'VirtualizedLists should never be nested inside plain ScrollView'
+  'Cache data may be lost when replacing'
 ])
 
 import App from './app/index'
